Update to 0.3.8.dev0.
Updating from the current master branch as there have been some bug
fixes since the release of 0.3.7.

Note that the python-gnupg is now maintained publicly on BitBucket
with a GitHub mirror here:

    https://github.com/vsajip/python-gnupg

Also note that a rival project forked from version 0.3.2 is now
operating with their version number increased to version 1.3.x and
they have *not* changed the module name of gnupg so installing it will
result in existing code breaking.  They also do not ship their code
under a license compatible with the Supybot license (they opted for
the GPLv3 and initially the Affero version rather than a license
conducive to being used as a library; ref. all the legal problems
relating to dynamic linking and python modules).